src/test/java/seedu/address/model/tag/TagTest.java
package seedu.address.model.tag;

import static seedu.address.testutil.Assert.assertThrows;

import org.junit.jupiter.api.Test;

public class TagTest {

    @Test
    public void constructor_nullTagName_throwsNullPointerException() {
        assertThrows(NullPointerException.class, () -> new Tag(null, TagStatus.INCOMPLETE_GOOD));
    }

    @Test
    public void constructor_nullTagStatus_throwsNullPointerException() {
        assertThrows(NullPointerException.class, () -> new Tag("Test", null));
    }

    @Test
    public void constructor_invalidTagName_throwsIllegalArgumentException() {
        String invalidTagName = "";
        assertThrows(IllegalArgumentException.class, () -> new Tag(invalidTagName, TagStatus.INCOMPLETE_GOOD));
    }

    @Test
    public void isValidTagName() {
        // null tag name
        assertThrows(NullPointerException.class, () -> Tag.isValidTagName(null));
    }

}
